ingestion-service/requirements.txt
boto3
python-dotenv
pypdf
docx2txt
pdfplumber
pika
transformers